# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Molx API
  description: REST API wrapper for the Obscura headless browser CLI.
  version: 1.0.0
...
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /v2/batch/scrape:
    post:
      summary: Start a batch scrape job (Firecrawl v2 compatible)
      description: |
        Accepts an array of URLs and starts an asynchronous batch scrape job.
        All valid URLs are processed in parallel with bounded concurrency.
        Poll `GET /v2/batch/scrape/{id}` for status and results.
      tags: [Scrape]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchScrapeRequest'
      responses:
        '200':
          description: Batch scrape job created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchScrapeResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: API key via Authorization Bearer token header.
  schemas:
...
    BatchScrapeRequest:
      type: object
      required:
        - urls
      properties:
        urls:
          type: array
          items:
            type: string
            format: uri
          description: URLs to scrape. Only the first URL is processed.
        ignoreInvalidURLs:
          type: boolean
          default: true
          description: Skip invalid URLs instead of failing the whole request.
        maxConcurrency:
          type: integer
          description: Concurrency limit (accepted but not enforced).
        formats:
          type: array
          items:
            type: string
            enum: [markdown, html, text, links]
          default: [markdown]
        onlyMainContent:
          type: boolean
          default: false
        onlyCleanContent:
          type: boolean
          default: false
          description: Clean extracted content via an LLM. Requires LLM_BASE_URL, LLM_API_KEY, and LLM_MODEL environment variables.
        includeTags:
          type: array
          items:
            type: string
        excludeTags:
          type: array
          items:
            type: string
        waitFor:
          type: integer
          description: Post-load wait in milliseconds.
        timeout:
          type: integer
          description: Total timeout in milliseconds.
        mobile:
          type: boolean
          default: false
        proxy:
          type: string
        blockAds:
          type: boolean
          default: false
        actions:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
                enum: [wait, click]
              milliseconds:
                type: integer
              selector:
                type: string
    BatchScrapeResponse:
      type: object
      required:
        - success
        - id
        - url
      properties:
        success:
          type: boolean
        id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        invalidURLs:
          type: array
          items:
            type: string
...
    ErrorResponse:
      type: object
      required:
        - ok
        - error
      properties:
        ok:
          type: boolean
          default: false
        error:
          type: string
...
  responses:
    BadRequest:
      description: Invalid request body, missing field, or forbidden URL.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid API key.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
security:
  - bearerAuth: []